% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{events_plot}
\alias{events_plot}
\title{Plot the output of eventsCount}
\usage{
events_plot(eventsCount.list, origin = NULL, period.length = NULL, E,
  use.letters = TRUE, call.mfrow = TRUE, ...)
}
\arguments{
\item{eventsCount.list}{A list of outputs of eventsCount().}

\item{origin}{Date of the beginning of the trial.}

\item{period.length}{Length of the time periods.}

\item{E}{Total number of events to achieve.}

\item{use.letters}{Index all elements in the list with a letter when plotting.}

\item{call.mfrow}{Wether to display all plots in the same window.}

\item{...}{Other arguments passed to par().}
}
\value{
Just produces one or more plots.
}
\description{
Plots the cummulative number of events observed and expected using the survival functions given to eventsCount().
}

