% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{eventsCount}
\alias{eventsCount}
\title{Total number of expected events}
\usage{
eventsCount(data = NULL, time.inclusion = NULL, time.end = NULL,
  time.followup = NULL, delta, abandon, new.inclusions = NULL,
  origin = NULL, period.length, past = FALSE, boot = TRUE,
  boot.samples = 500, varnames = getOption("RCT_names_match"), surv.object,
  how.many = 1000, FUN = NULL)
}
\arguments{
\item{data}{Defaults to NULL. Data frame with the inclusion times (named as inclusion), follow-up times (y) or ending times (end),
withdrawal indicator (abandon) and censoring indicator (delta). You can provide time.inclusion, delta,
abandon and time.end or time.followup instead.}

\item{time.inclusion}{Ignored if data is not NULL. Vector with times of inclusion (numeric or date).}

\item{time.end}{Ignored if data is not NULL. Vector with the times of end of follow-up (numeric or date). One
of time.followup or time.end must be supplied (if data is not null)}

\item{time.followup}{Ignored if data is not NULL. Numeric vector with the total time of follow-up. One
of time.followup or time.end must be supplied (if data is not null)}

\item{delta}{Ignored if data is not NULL. Censoring indicator: 1 if the event is observed, 0 otherwise}

\item{abandon}{Ignored if data is not NULL. Withdrawal indicator: 1 if the patient is a withdrawal, abandons,
or dies for other causes (etc). 0 otherwise.}

\item{new.inclusions}{Numeric vector with the new patients included in each future time period.}

\item{origin}{Date that marks the first included patient in the RCT.}

\item{period.length}{Length of each time period.}

\item{past}{Logical. Calculate the expected number of events in each past time periods in the data provided?}

\item{FUN}{Survival function for calculating probabilities of event in each period.}
}
\value{
A list with
\item{events}{Numeric with the cummulative number of events in each period (past and future).}
\item{percent.events}{Numeric with the percentage of events in the sample (sample size is modified in every period).}
\item{new.inclusions}{Argument new.inclusions.}
\item{last.time}{Highest time in the data: max(inclusion + followup).}
\item{data}{Data frame constructed from the arguments provided (or just the data argument with some modifications, if provided).}
\item{survFun}{Argument FUN.}
\item{origin}{Argument origin or the origin assumed for the data (first inclusion, when the RCT started).}
}
\description{
Calculates the total number of expected events (past and/or future) in a RCT in each time period.
}
\seealso{
\code{\link{survs_plot}}, \code{\link{events_plot}}, \code{\link{inclusionsCount}}
}

