% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{inclusions_plot}
\alias{inclusions_plot}
\title{Plot the output of inclusionsCount}
\usage{
inclusions_plot(inclusionsCount.list, origin = NULL, period.length = NULL,
  N, use.letters = TRUE, call.mfrow = TRUE, ...)
}
\arguments{
\item{inclusionsCount.list}{A list of outputs of inclusionsCount().}

\item{origin}{Date of the beginning of the trial.}

\item{period.length}{Length of the time periods.}

\item{N}{Total number of included patients to achieve.}

\item{use.letters}{Index all elements in the list with a letter when plotting.}

\item{call.mfrow}{Wether to display all plots in the same window.}

\item{...}{Other arguments passed to par().}
}
\value{
Just produces one or more plots.
}
\description{
Produces two plots: one with the counts of each time period and the predicted ones, and the other with the cummulative counts.
}

