% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{inclusionsCount}
\alias{inclusionsCount}
\title{Predicting the number of inclusions in a time period}
\usage{
inclusionsCount(time.inclusion = NULL, period.inclusion = NULL,
  model.args = list("glm", formula = counts ~ time.periods +
  I(time.periods^2), family = poisson), do.before = NULL, do.after = NULL,
  predict.next = 10, boot = TRUE, boot.samples = 500, period.length,
  origin = NULL, numeric = TRUE)
}
\arguments{
\item{time.inclusion}{Inclusion times for the patients in the study.}

\item{period.inclusion}{Period of inclusion of the patients in the study. Either this or time.inlcusion has to be passed.}

\item{model.args}{List with the arguments needed for the model, called via do.call(model.args[[1]], model.args[-1]).}

\item{do.before, }{do.after Expressions to be executed before and after fitting the model.}

\item{predict.next}{Number of periods of which to predict the inclusions.}

\item{boot}{Calculate or not bootstrap CI intervals.}

\item{boot.samples}{Number of samples for the CI intervals.}

\item{period.length, }{origin, numeric Arguments passed to as.period if time.inclusion is used instead of period.inclusion.}
}
\value{
A list with the counts predicted in each period using the model given, the cummulative counts, bootstrap CI for the counts and cummulative counts
if asked, the data used and the last period.
}
\description{
Predict the number of total inclusions in future time periods using the passed model for the counts.
}
\details{
Data can be passed using time.inclusion with the days elapsed from the beginning of the RCT until the inclusion of each patient, or period.inclusion
with the period of the study that the patient was recruited. In the first case peroid.length, origin and/or numeric can be passed also.
For fitting a poisson glm for instance, set model.args as list("glm", formula, family = poisson). Then do.before and do.after can be used to modify
or change the data as needed. substitute(df2$counts) can be used to get the data.
}

