% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{period_adj}
\alias{period_adj}
\title{Removing last time period}
\usage{
period_adj(data, period.length)
}
\arguments{
\item{data}{Data to remove last time period. A data frame with variables inclusion, y, abandon and delta at least. Other variables won't be modified.}

\item{period.length}{Length of each period.}
}
\value{
The same data frame but without the patients included in the last period of time (removed), and some modified delta's.
}
\description{
Removes last time period from the data, and changes those patients that had event in this period. Useful in case that
the last time period has not finished.
}

