% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{periods_plot}
\alias{periods_plot}
\title{Plot histogram for different lengths of the time periods}
\usage{
periods_plot(period.length, time.inclusion, letters = TRUE, ...)
}
\arguments{
\item{period.length}{Vector with the lengths for the periods. One plot is produced for each.}

\item{time.inclusion}{Vector with the inclusion times.}

\item{letters}{Whether to plot alphabet letters top-left corner of each plot.}

\item{...}{Further arguments passed to par.}
}
\value{
No output is produced. Only a plot.
}
\description{
Decide which calendar partition suits better the data by a visual inspection
}

