% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRCT.R
\docType{package}
\name{predRCT}
\alias{predRCT}
\alias{predRCT-package}
\title{predRCT}
\description{
Functions used in "Predicting the remaining time to achieve the required 
number of needed patients or events in an ongoing clinical trial", by Gerard Castellà Canals.
}

