% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{probAB}
\alias{probAB}
\title{Calculate proabilities of event}
\usage{
probAB(patients = TRUE, data, start, end, last.time, FUN)
}
\arguments{
\item{patients}{Subset of indices for which to calculate the probabilities}

\item{data}{Dataset with the times of inclussion, withdrawals, censoring indicator, time of follow-up, ...}

\item{start, }{end Start or end of the interval of which to calculate the probabilities.}

\item{last.time}{Number of days elapsed since the beginning of the trial.}

\item{FUN}{survival function used for calculating the probabilities. Usually output from survFUN().}
}
\value{
A vector with the probabilities with the same length as nrow(data[patients, , drop = FALSE]).
}
\description{
Calculate proabilities of event for a set/subset of patients in a time interval, using a given survival function.
}

