% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{survFUN}
\alias{survFUN}
\title{Generate Survival Function}
\usage{
survFUN(object, period.length, how.many = 1000)
}
\arguments{
\item{object}{survfit or survreg object.}

\item{period.length}{Length of each time period.}

\item{how.many}{How many quantiles you want for survreg case.}
}
\value{
A function with one parameter: the vector of times to calculate the survival probabilities.
}
\description{
Taking survfit or survreg output to create the corresponding survival function.
}
\seealso{
\code{\link{survs_plot}}, \code{\link{eventsCount}}.
}

