% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{survs_plot}
\alias{survs_plot}
\title{Plot survival functions}
\usage{
survs_plot(surv.funs, survfit.object, period.length, xlim = 2 *
  max(summary(survfit.object)$time), call.par = TRUE, letters = call.par,
  rev = TRUE, ...)
}
\arguments{
\item{surv.funs}{List of functions returned by \code{survFUN}.}

\item{survfit.object}{Object returned by survfit.}

\item{period.length}{Length of each period.}

\item{xlim}{Maximum xlim for the plot.}

\item{call.par}{Call \code{par()} with mfrow for plotting all graphics in the same figure.}

\item{letters}{Whether to plot alphabet letters top-left corner of each plot.}

\item{rev}{Defaults to TRUE. See argument rev in \code{statTools::descomp}. It is reverses the vector passed to \code{mfrow} argument.}

\item{...}{Further arguments passed to par.}
}
\value{
No output is produced. Only a plot.
}
\description{
Plot survival functions returned by \code{survFUN()}. KM estimation of the survival function is also plotted (survfit.object).
}

