% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2.R
\name{as.period}
\alias{as.period}
\title{Time Period}
\usage{
as.period(times, period.length, numeric = TRUE)
}
\arguments{
\item{times}{Vector of times to transform to periods.}

\item{period.length}{Length of each time period.}

\item{numeric}{Logical. Return the periods as a numeric? Defaults to TRUE}
}
\value{
Returns a character (or numeric when \code{numeric = TRUE}) vector of length values with
the corresponding time periods for each time.
}
\description{
Transform time unit to period of time
}
\details{
Time periods are closed on the right, and 0 is included in the first one.
Period labels, when \code{numeric} is \code{FALSE}, come from \code{cut}
}

